requirements.txt
python-dotenv
torch
sentencepiece
tiktoken
safetensors
requests
thop
tqdm